chore: Reset view state in storage
